+++ src/lib.rs
@@ -583,9 +583,10 @@ pub mod prelude {
     pub use crate::structures::element::Elements;
     pub use crate::structures::gridmap::GridMap;
     pub use crate::structures::iterators::{
-        AtomIterator, AtomIteratorWithBox, AtomPairIterator, FilterAtomIterator, MoleculeIterator,
-        MutAtomIterator, MutAtomIteratorWithBox, MutAtomPairIterator, MutFilterAtomIterator,
-        MutMoleculeIterator, OwnedAtomIterator, OwnedMutAtomIterator,
+        AtomIterator, AtomIteratorWithBox, AtomPairIterator, FilterAtomIterator,
+        IntersectionAtomIterator, MoleculeIterator, MutAtomIterator, MutAtomIteratorWithBox,
+        MutAtomPairIterator, MutFilterAtomIterator, MutMoleculeIterator, OrderedAtomIterator,
+        OwnedAtomIterator, OwnedMutAtomIterator, UnionAtomIterator,
     };
     pub use crate::structures::shape::{Cylinder, Rectangular, Shape, Sphere, TriangularPrism};
     pub use crate::structures::simbox::SimBox;