+++ changelog.md
@@ -33,12 +33,14 @@
 - **Breaking change**: `System::traj_iter_map_reduce` now requires the user to provide the initial `Data` structure. `Data` structure no longer needs to implement `Default`, but needs to implement `Clone`.
 - `System::traj_iter_map_reduce` now properly propagates errors and failures from the individual threads into the master-thread `ProgressPrinter`.
 
-#### Atom iterators
+#### Changes to atom iterators
 - **Breaking change:** Renamed `MasterAtomIterator` trait to `AtomIteratorWithBox` and `MasterMutAtomIterator` trait to `MutAtomIteratorWithBox`.
+- Added a new trait `OrderedAtomIterator` which is implemented by all immutable iterators which yield atoms in the order in which they are present in the molecular system. Analogously, implemented `MutOrderedAtomIterator`.
+- Iterator unions and intersections can be constructed for iterators implementing `OrderedAtomIterator`(`MutOrderedAtomIterator`) using `(Mut)OrderedAtomIterator::union` and `(Mut)OrderedAtomIterator::intersection` respectively.
 - Added `AtomPairIterator` and `MutAtomPairIterator` to enable iteration over pairs of atoms. These iterators currently support iterating over molecule bonds and can be created via `System::molecule_bonds_iter` and `System::molecule_bonds_iter_mut` for immutable and mutable access, respectively.
 - Note: The design of these iterators and the structure of atom pairs are subject to future refinement and potential expansion in functionality.
 
-#### Bug fixes and minor changes
+#### Bug fixes and other changes
 - Introduced `System::make_group_whole` for making a group "whole" in the system (similar to `System::make_molecules_whole`).
 - **Breaking change:** `System::group_all_distances` now returns ndarray's `Array2` instead of a vector of vectors.
 - `FileType` is now part of the prelude.